move dashboard components to project app
