+++ meinberlin/apps/projects/dashboard.py
@@ -1,0 +1,53 @@
+from django.urls import reverse
+from django.utils.translation import ugettext_lazy as _
+
+from meinberlin.apps.dashboard2 import DashboardComponent
+from meinberlin.apps.dashboard2 import components
+
+from . import views
+
+
+class ParticipantsComponent(DashboardComponent):
+    identifier = 'participants'
+    weight = 30
+    label = _('Participants')
+
+    def is_effective(self, project):
+        return project.is_private
+
+    def get_base_url(self, project):
+        return reverse('a4dashboard:dashboard-participants-edit', kwargs={
+            'project_slug': project.slug
+        })
+
+    def get_urls(self):
+        return [(
+            r'^projects/(?P<project_slug>[-\w_]+)/participants/$',
+            views.DashboardProjectParticipantsView.as_view(component=self),
+            'dashboard-participants-edit'
+        )]
+
+
+class ModeratorsComponent(DashboardComponent):
+    identifier = 'moderators'
+    weight = 32
+    label = _('Moderators')
+
+    def is_effective(self, project):
+        return True
+
+    def get_base_url(self, project):
+        return reverse('a4dashboard:dashboard-moderators-edit', kwargs={
+            'project_slug': project.slug
+        })
+
+    def get_urls(self):
+        return [(
+            r'^projects/(?P<project_slug>[-\w_]+)/moderators/$',
+            views.DashboardProjectModeratorsView.as_view(component=self),
+            'dashboard-moderators-edit'
+        )]
+
+
+components.register_project(ModeratorsComponent())
+components.register_project(ParticipantsComponent())
